Implement simple circle collision handling
Circle is now capable of moving as close as possible to the line
if it is colliding with a single line. There are still some issues
that can be seen if the angle between the line and the velocity
vector is really small and when the circle moves along the canvas
borders and collides with a line.